# openapi.yaml
openapi: 3.0.3
info:
  title: Omniscience API
  description: REST API for managing educational data including courses, classes, students, and attendance records
  version: 1.0.0
  contact:
    name: Omniscience API Team
servers:
  - url: http://localhost:8080
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Supabase authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    GradeScale:
      type: object
      additionalProperties:
        type: object
        properties:
          min:
            type: number
            format: float
          max:
            type: number
            format: float
      example:
        A: { min: 90, max: 100 }
        B: { min: 80, max: 89 }
        C: { min: 70, max: 79 }
        D: { min: 60, max: 69 }
        F: { min: 0, max: 59 }
    GradingFormula:
      type: object
      properties:
        id:
          type: integer
          format: uint
        course_id:
          type: integer
          format: uint
        formula_type:
          type: string
          enum: [WEIGHTED_AVERAGE, POINTS_BASED, PASS_FAIL, CUSTOM]
          description: Type of grading formula
        passing_percentage:
          type: number
          format: float
          description: Minimum percentage required to pass
        grading_scale:
          $ref: '#/components/schemas/GradeScale'
        formula_config:
          type: object
          description: Additional configuration for custom formulas
        is_active:
          type: boolean
          description: Whether this formula is currently active
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - course_id
        - formula_type
    GradingFormulaRequest:
      type: object
      properties:
        course_id:
          type: integer
          format: uint
        formula_type:
          type: string
          enum: [WEIGHTED_AVERAGE, POINTS_BASED, PASS_FAIL, CUSTOM]
        passing_percentage:
          type: number
          format: float
        grading_scale:
          $ref: '#/components/schemas/GradeScale'
        formula_config:
          type: object
      required:
        - course_id
        - formula_type
...
    Grade:
      type: object
      properties:
        id:
          type: integer
          format: uint
        evaluation_item_id:
          type: integer
          format: uint
        registration_id:
          type: integer
          format: uint
        score:
          type: number
          format: float
        percentage:
          type: number
          format: float
          description: Automatically calculated percentage score
        letter_grade:
          type: string
          description: Automatically assigned letter grade
        is_excused:
          type: boolean
        is_late:
          type: boolean
        late_penalty:
          type: number
          format: float
        comments:
          type: string
        graded_by:
          type: string
        graded_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [DRAFT, PUBLISHED]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - evaluation_item_id
        - registration_id
    GradeRequest:
      type: object
      properties:
        evaluation_item_id:
          type: integer
          format: uint
        registration_id:
          type: integer
          format: uint
        score:
          type: number
          format: float
        is_excused:
          type: boolean
          default: false
        is_late:
          type: boolean
          default: false
        late_penalty:
          type: number
          format: float
        comments:
          type: string
        graded_by:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED]
          default: DRAFT
      required:
        - evaluation_item_id
        - registration_id
    BatchGradeItem:
      type: object
      properties:
        registration_id:
          type: integer
          format: uint
        score:
          type: number
          format: float
        is_excused:
          type: boolean
        is_late:
          type: boolean
        comments:
          type: string
      required:
        - registration_id
    BatchGradeRequest:
      type: object
      properties:
        evaluation_item_id:
          type: integer
          format: uint
        grades:
          type: array
          items:
            $ref: '#/components/schemas/BatchGradeItem'
      required:
        - evaluation_item_id
        - grades
    BatchGradeResponse:
      type: object
      properties:
        message:
          type: string
        created:
          type: integer
        updated:
          type: integer
        grades:
          type: array
          items:
            $ref: '#/components/schemas/Grade'
    PublishGradesResponse:
      type: object
      properties:
        message:
          type: string
        published:
          type: integer
...
security:
  - bearerAuth: []
paths:
...
  /grading-formulas:
    get:
      summary: Get grading formula
      description: Get the active grading formula for a course
      operationId: getGradingFormula
      tags:
        - Grading Formulas
      security:
        - bearerAuth: []
      parameters:
        - name: course_id
          in: query
          description: ID of the course
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradingFormula'
        '404':
          description: Grading formula not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create grading formula
      description: Create a new grading formula for a course
      operationId: createGradingFormula
      tags:
        - Grading Formulas
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradingFormulaRequest'
      responses:
        '201':
          description: Formula created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradingFormula'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /grading-formulas/{id}:
    put:
      summary: Update grading formula
      description: Update an existing grading formula
      operationId: updateGradingFormula
      tags:
        - Grading Formulas
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Formula ID
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradingFormulaRequest'
      responses:
        '200':
          description: Formula updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradingFormula'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Formula not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /grades:
    get:
      summary: List grades
      description: List grades with optional filters
      operationId: listGrades
      tags:
        - Grades
      security:
        - bearerAuth: []
      parameters:
        - name: evaluation_item_id
          in: query
          description: Filter by evaluation item
          required: false
          schema:
            type: integer
            format: uint
        - name: registration_id
          in: query
          description: Filter by student registration
          required: false
          schema:
            type: integer
            format: uint
        - name: student_class_id
          in: query
          description: Filter by class
          required: false
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Grade'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create grade
      description: Create a single grade
      operationId: createGrade
      tags:
        - Grades
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeRequest'
      responses:
        '201':
          description: Grade created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /grades/{id}:
    get:
      summary: Get grade
      description: Get a specific grade
      operationId: getGrade
      tags:
        - Grades
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Grade ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '404':
          description: Grade not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update grade
      description: Update a grade. Creates audit history record.
      operationId: updateGrade
      tags:
        - Grades
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Grade ID
          required: true
          schema:
            type: integer
            format: uint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeRequest'
      responses:
        '200':
          description: Grade updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grade'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Grade not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /grades/batch:
    post:
      summary: Batch create grades
      description: Create multiple grades at once for an evaluation item
      operationId: batchCreateGrades
      tags:
        - Grades
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGradeRequest'
      responses:
        '201':
          description: Grades created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchGradeResponse'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /grades/publish/{evaluation_item_id}:
    post:
      summary: Publish grades
      description: Publish all draft grades for an evaluation item
      operationId: publishGrades
      tags:
        - Grades
      security:
        - bearerAuth: []
      parameters:
        - name: evaluation_item_id
          in: path
          description: Evaluation Item ID
          required: true
          schema:
            type: integer
            format: uint
      responses:
        '200':
          description: Grades published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishGradesResponse'
        '404':
          description: Evaluation item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...